Fix to support Expo 47
Applied fixes from coffeOwl shared here: https://github.com/expo/expo/issues/18784#issuecomment-1333726866

fix for android build

Added the watermelon db package to the main activity init package list as it appears to be needed (i.e. auto-linking is not working for me). Also added an escape for the simulator architecture exclusion to the config option for the expo plugin.